`data-use-parent` attribute option added. Fixes #14
When a image has the `data-use-parent` attribute the script will use the
width of it's parent instead of the with of the viewport.

This way the images can also be made responsive when used in for example a grid layout.

Next to this I made some other changes to simplify things a bit:

* onLoad and onResize all img elements on the page will be processed
* makeImagesResponsive is however available to the global scope and
accepts an argument to specify a specific array of image elements to process